backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# Cache & Background Tasks
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Real-time Communication
channels==4.0.0
channels-redis==4.1.0

# File & Media Handling
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0

# API Documentation
drf-spectacular==0.26.5

# Utilities
python-decouple==3.8
qrcode==7.4.2
geopy==2.4.1

# Development
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest-django==4.7.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
